refactor(style-modal): renaming style-modal to style-modal-drawing
