+++ packages/geo/src/lib/draw/draw/draw.component.html
@@ -169,7 +169,7 @@
     mat-raised-button
     color="primary"
     (click)="openStyleModalDialog()"
-    [matTooltip]="'igo.geo.style.styleModalTooltip' | translate">
+    [matTooltip]="'igo.geo.style.styleModalTooltipSelected' | translate">
     {{'igo.geo.style.styleModal' | translate}}
     <mat-icon
       class="style-icon"
