+++ packages/geo/src/lib/style/style-modal/drawing/index.ts
@@ -1,0 +1,1 @@
+export * from './style-modal-drawing.component';
