+++ packages/geo/src/lib/style/style-modal/drawing/style-modal-drawing.component.scss
