+++ packages/geo/src/lib/style/style-modal/index.ts
@@ -1,1 +1,2 @@
-export * from './style-modal.component';
+export * from './drawing';
+export * from './shared';
