+++ packages/geo/src/lib/style/style-modal/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './style-modal.interface';
