+++ packages/geo/src/locale/en.geo.json
@@ -262,7 +262,15 @@
       },
       "style": {
         "styleModal": "Edit style",
-        "styleModalTooltip": "Edit the style of the selected entities "
+        "styleModalTooltipSelected": "Edit the style of the selected entities",
+        "colorPicker": "Click to open the color selector",
+        "cancel": "Cancel",
+        "fill": "Fill",
+        "stroke": "Stroke",    
+        "fontSize": "Font size",
+        "fontStyle": "Font style",
+        "offsetX": "Horizontal offset",
+        "offsetY": "Vertical offset"
       },
       "download": {
         "action": "Download data",
@@ -685,12 +693,10 @@
         "abort": "Abort current drawing",
         "cancel": "Cancel",
         "Circle": "Circle",
-        "colorPicker": "Click to open the color selector",
         "dialogInstruction": "Enter your drawing's label",
         "dialogTitle": "Label",
         "delete": "Remove Drawings from map",
         "drawing": "Drawings",
-        "fill": "Fill",
         "finish": "Finish current drawing",
         "freehandMode": "Freehand drawing",
         "icon": "Icons",
@@ -704,14 +710,9 @@
         "Polygon": "Polygon",
         "radius": "Radius",
         "shortcuts": "Keyboard shortcuts",
-        "stroke": "Stroke",
         "styleSelect": "Select style",
         "toggleMapTooltips": "Show map tooltips",
         "undo": "Remove last vertex",
-        "fontSize": "Font size",
-        "fontStyle": "Font style",
-        "offsetX": "Horizontal offset",
-        "offsetY": "Vertical offset",
         "layer": {
           "title": "Title",
           "createNewLayer": "Create new drawing layer",
