+++ packages/geo/src/locale/fr.geo.json
@@ -261,7 +261,15 @@
       },
       "style": {
         "styleModal": "Modifier le style",
-        "styleModalTooltip": "Éditer le style des entités sélectionnées"
+        "styleModalTooltipSelected": "Éditer le style des entités sélectionnées",
+        "colorPicker": "Cliquer pour ouvrir le sélecteur de couleurs",
+        "cancel": "Annuler",
+        "fill": "Remplissage",
+        "stroke": "Contour",
+        "fontSize": "Taille de police",
+        "fontStyle": "Style de police",
+        "offsetX": "Décalage horizontal",
+        "offsetY": "Décalage vertical"
       },
       "download": {
         "action": "Télécharger les données associées",
@@ -685,12 +693,10 @@
         "abort": "Annuler le dessin en cours",
         "cancel": "Annuler",
         "Circle": "Cercle",
-        "colorPicker": "Cliquer pour ouvrir le sélecteur de couleurs",
         "dialogInstruction": "Veuillez entrer l'étiquette de votre dessin",
         "dialogTitle": "Étiquette",
         "delete": "Retirer les dessins de la carte",
         "drawing": "Dessins",
-        "fill": "Remplissage",
         "finish": "Terminer le dessin en cours",
         "freehandMode": "Dessiner à main levée",
         "icon": "Icônes",
@@ -704,14 +710,9 @@
         "Polygon": "Polygone",
         "radius": "Rayon",
         "shortcuts": "Raccourcis clavier",
-        "stroke": "Contour",
         "styleSelect": "Sélectionner le style",
         "toggleMapTooltips": "Afficher les étiquettes",
         "undo": "Retirer le dernier sommet",
-        "fontSize": "Taille de police",
-        "fontStyle": "Style de police",
-        "offsetX": "Décalage horizontal",
-        "offsetY": "Décalage vertical",
         "layer": {
           "title": "Titre",
           "createNewLayer": "Créer une nouvelle couche de dessin",
